% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopri5.R
\name{dopri5}
\alias{dopri5}
\title{Integrate ODE/DDE with dopri5}
\usage{
dopri5(y, times, func, parms, ..., n_out = 0L, output = NULL,
  rtol = 1e-06, atol = 1e-06, n_history = 0, keep_history = n_history >
  0, dllname = "", parms_are_real = TRUE)
}
\arguments{
\item{y}{Initial conditions for the integration}

\item{times}{Times where output is needed.  Unlike \code{deSolve}
we won't actually stop at these times, but instead interpolate
back to get the result.}

\item{func}{Function to integrate.  Can be an R function of
arguments \code{t, y, parms}, returning a numeric vector, or it
can be the name or address of a C function with arguments
\code{size_t n, double t, double *y, double *dydt, void *data}.}

\item{parms}{Parameters to pass through to the derivatives.}

\item{...}{Dummy arguments - nothing is allowed here, but this
means that all further arguments \emph{must} be specified by
name (not order) so I can easily reorder them later on.}

\item{n_out}{Number of "output" variables (not differential
equation variables) to compute via the routine \code{output}.}

\item{output}{The output routine; either an R function taking
arguments \code{t, y, parms} or the name/address of a C function
taking arguments \code{size_t n, double t, const double *y,
size_t n_out, double *out, void *data}.}

\item{rtol}{The per-step relative tolerance.  The total accuracy
will be less than this.}

\item{atol}{The per-step absolute tolerance.}

\item{n_history}{Number of history points to retain.  This needs
to be greater than zero for delay differential equations to
work.  Alternatively, this may be greater than zero to return
model outputs that can be inspected later.}

\item{keep_history}{Logical indicating if history should be
retained or discarded.  By default, history is retained}

\item{dllname}{Name of the shared library (without extension) to
find the function \code{func} (and \code{output} if given) in
the case where \code{func} refers to compiled function.}

\item{parms_are_real}{Logical, indicating if \code{parms} should
be treated as vector of doubles by \code{func} (when it is a
compiled function).  If \code{TRUE} (the default), then
\code{REAL(parms)}, which is \code{double*} is passed through.
If \code{FALSE} then if \code{params} is an externalptr type
(\code{EXTPTRSXP}) we pass through the result of
\code{R_ExternalPtrAddr}, otherwise we pass \code{params}
through unmodified as a \code{SEXP}.  In the last case, in your
target function you will need to include \code{<Rinternals.h>},
cast to \code{SEXP} and then pull it apart using the R API (or
Rcpp).}
}
\value{
At present the return value is transposed relative to
  deSolve.  This might change in future.
}
\description{
Integrate an ODE or DDE with dopri5.
}

